refactor: isolate sandbox code from main CLI paths
Extract all sandbox-mode code into dedicated modules so non-sandbox
users are completely unaffected. Sandbox is now an experimental
command group (`defenseclaw sandbox init|setup`) rather than flags
on the main init/setup commands.

- Move sandbox init helpers to cmd_init_sandbox.py
- Move sandbox setup command to cmd_setup_sandbox.py
- Create cmd_sandbox.py group marked [experimental]
- Remove --sandbox flag from `defenseclaw init`
- Remove sandbox subcommand from `defenseclaw setup`
- Split sandbox/firewall OPA data into data-sandbox.json
- Remove unconditional sandbox_home default from config
- Gate _preserve_ownership on root-only (guardrail.py)
- Comment out UpdateSandboxPolicy in watcher (TODO)
- Remove /policy/evaluate/sandbox API endpoint
- Remove update_sandbox_policy from policy defaults
- Add unit tests for mergeSupplementalData
- Restore Makefile to origin/main

Made-with: Cursor
