+++ Makefile
@@ -14,7 +14,7 @@ DIST_DIR    := dist
 .PHONY: build install dev-install pycli dev-pycli gateway gateway-cross gateway-run start gateway-install \
         plugin plugin-install test cli-test cli-test-cov gateway-test go-test-cov \
         test-verbose test-file lint py-lint go-lint ts-test rego-test clean \
-        dist dist-cli dist-gateway dist-plugin dist-sandbox dist-test dist-checksums dist-clean
+        dist dist-cli dist-gateway dist-plugin dist-checksums dist-clean
 
 # ---------------------------------------------------------------------------
 # Aggregate targets
@@ -40,17 +40,6 @@ install: pycli gateway-install plugin-install
 	@echo "  source $(VENV)/bin/activate"
 	@echo "  defenseclaw init"
 	@echo "  defenseclaw setup guardrail   # configure LLM guardrail"
-	@echo ""
-	@if [ "$$(uname -s)" = "Linux" ]; then \
-		echo "Sandbox mode (Linux):"; \
-		echo "  defenseclaw init --sandbox          # create sandbox user + directories"; \
-		echo "  defenseclaw setup sandbox            # configure networking + systemd"; \
-		echo "  scripts/install-openshell-sandbox.sh  # install openshell-sandbox binary"; \
-	else \
-		echo "Sandbox mode (Linux only):"; \
-		echo "  On a Linux host, use 'defenseclaw init --sandbox' to set up"; \
-		echo "  openshell-sandbox standalone mode with network isolation."; \
-	fi
 
 # ---------------------------------------------------------------------------
 # Individual build targets
@@ -64,8 +53,6 @@ pycli:
 	@find cli/ -type d -name __pycache__ -exec rm -rf {} + 2>/dev/null || true
 	uv venv $(VENV) --python 3.12
 	uv pip install -e . --python $(VENV)/bin/python
-	@mkdir -p $(INSTALL_DIR)
-	@ln -sf $(CURDIR)/$(VENV)/bin/litellm $(INSTALL_DIR)/litellm 2>/dev/null || true
 
 dev-pycli: pycli
 	uv pip install --group dev --python $(VENV)/bin/python
@@ -191,7 +178,7 @@ go-lint:
 # Distribution targets — build release artifacts into dist/
 # ---------------------------------------------------------------------------
 
-dist: dist-cli dist-gateway dist-plugin dist-sandbox dist-test dist-checksums
+dist: dist-cli dist-gateway dist-plugin dist-checksums
 	@echo ""
 	@echo "Release artifacts:"
 	@ls -lh $(DIST_DIR)/
@@ -209,17 +196,12 @@ dist-cli: _bundle-data
 
 _bundle-data:
 	@mkdir -p cli/defenseclaw/_data/policies/rego
-	@mkdir -p cli/defenseclaw/_data/policies/openshell
-	@mkdir -p cli/defenseclaw/_data/scripts
 	@mkdir -p cli/defenseclaw/_data/skills
 	@rm -rf cli/defenseclaw/_data/splunk_local_bridge
 	cp policies/rego/*.rego cli/defenseclaw/_data/policies/rego/
 	rm -f cli/defenseclaw/_data/policies/rego/*_test.rego
 	cp policies/rego/data.json cli/defenseclaw/_data/policies/rego/
 	cp policies/*.yaml cli/defenseclaw/_data/policies/
-	cp policies/openshell/*.rego cli/defenseclaw/_data/policies/openshell/
-	cp policies/openshell/*.yaml cli/defenseclaw/_data/policies/openshell/
-	cp scripts/install-openshell-sandbox.sh cli/defenseclaw/_data/scripts/
 	cp -r skills/codeguard cli/defenseclaw/_data/skills/
 	cp -r bundles/splunk_local_bridge cli/defenseclaw/_data/
 
@@ -245,35 +227,14 @@ dist-plugin: plugin
 		done)
 	@echo "Plugin tarball built"
 
-dist-sandbox:
-	@mkdir -p $(DIST_DIR)/sandbox/policies $(DIST_DIR)/sandbox/scripts
-	cp policies/openshell/*.rego $(DIST_DIR)/sandbox/policies/
-	cp policies/openshell/*.yaml $(DIST_DIR)/sandbox/policies/
-	cp scripts/install-openshell-sandbox.sh $(DIST_DIR)/sandbox/scripts/
-	chmod +x $(DIST_DIR)/sandbox/scripts/install-openshell-sandbox.sh
-	@echo "Sandbox artifacts copied to $(DIST_DIR)/sandbox/"
-
-dist-test:
-	@mkdir -p $(DIST_DIR)/test
-	cp scripts/test-litellm-proxy.py $(DIST_DIR)/test/
-	cp scripts/test-e2e-tool-block.sh $(DIST_DIR)/test/
-	cp scripts/test-e2e-sandbox-policy-diff.sh $(DIST_DIR)/test/ 2>/dev/null || true
-	cp scripts/test-e2e-cli.py $(DIST_DIR)/test/ 2>/dev/null || true
-	cp scripts/test-e2e-spark.sh $(DIST_DIR)/test/ 2>/dev/null || true
-	cp scripts/test-e2e-mac.sh $(DIST_DIR)/test/ 2>/dev/null || true
-	cp scripts/bundle-sandbox-test.sh $(DIST_DIR)/test/ 2>/dev/null || true
-	chmod +x $(DIST_DIR)/test/*.sh 2>/dev/null || true
-	@echo "Test scripts copied to $(DIST_DIR)/test/"
-
 dist-checksums:
 	@test -d $(DIST_DIR) || { echo "Run 'make dist' first"; exit 1; }
-	cd $(DIST_DIR) && find . -type f ! -name checksums.txt | sort | xargs shasum -a 256 > checksums.txt
+	cd $(DIST_DIR) && shasum -a 256 * > checksums.txt
 	@echo "Checksums written to $(DIST_DIR)/checksums.txt"
 
 dist-clean:
 	rm -rf $(DIST_DIR)
 	rm -rf cli/defenseclaw/_data
-	rm -rf sandbox-test-*
 
 clean:
 	rm -f $(GATEWAY) $(BINARY)-linux-* $(BINARY)-darwin-*
